requirements.txt
streamlit>=1.36.0
sqlite-utils
Pillow
streamlit-drawable-canvas
numpy